API Call works on the search JS files/page.
