+++ server.js
@@ -31,6 +31,7 @@ app.set("view engine", "handlebars");
 // // Routes
 require("./routes/apiRoutes")(app);
 require("./routes/htmlRoutes")(app);
+require("./routes/searchRoutes")(app);
 
 var syncOptions = { force: false };
